+++ tasks/titaev_m_avg_el_vector/seq/src/ops_seq.cpp
@@ -26,7 +26,7 @@ bool TitaevMElemVecsAvgSEQ::RunImpl() {
     sum += v;
   }
 
-  GetOutput() = sum / vec.size();
+  GetOutput() = sum / static_cast<double>(vec.size());
   return true;
 }
